fixing grey header in pdf
with an update to wkhtmltopdf version 0.12.6 the behavior regarding referencing
local files changed. It was disabled by default. The grey header comes from
such a file that is generated in the temp directory and can't be included. In
order to fix the issue the command line parameter --enable-local-file-access
needs to be used when calling wkhtmltopdf